% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday.R
\name{holiday}
\alias{holiday}
\title{A Holiday Function}
\usage{
holiday(christmas = TRUE)
}
\arguments{
\item{christmas}{Do you like holidays? Defaults to TRUE.}
}
\description{
This function allows you to decide what to do during Christmas Time
}
\examples{
holiday()
}
\keyword{holiday}
